% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesim-package.R
\docType{package}
\name{agesim}
\alias{agesim}
\title{Simulated age-structured populations}
\description{
Functions and data for simulating age-structured populations.
This package is experimental. The functionality is likely
to be absorbed into package 'demtech' when that is
written.
}
