% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lx_south.R
\docType{data}
\name{Lx_south}
\alias{Lx_south}
\title{Life table Lx values from the Soutn model life table}
\format{
An object of class \code{array} of dimension 15 x 2 x 25.
}
\source{
The values were obtained from the
\code{\link[demogR]{cdmlts}} function
in package \code{demogR}.
}
\usage{
Lx_south
}
\description{
The life table function Lx, by age, sex, and life table
level, for the South model life table.
}
\details{
For a description of life tables, and West model life tables,
see \url{http://demographicestimation.iussp.org}.
}
\seealso{
\code{\link{mx_south}}
}
\keyword{datasets}
