% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_account.R
\name{make_account}
\alias{make_account}
\title{Make an initial demographic account}
\usage{
make_account(expected_popn, mort_rates, fert_rates, time_start, time_end)
}
\arguments{
\item{expected_popn}{A \code{\link[dembase:Values-class]{Values}}
array giving expected population counts by age and sex. Typically
created using function \code{\link{make_expected_popn}}.}

\item{mort_rates}{A \code{\link[dembase:Values-class]{Values}}
array giving mortality rates by age and sex.}

\item{fert_rates}{A \code{\link[dembase:Values-class]{Values}}
array giving mean fertility rates by age and sex. Typically
generated using function \code{\link{make_stationary_fert_rates}}.}

\item{time_start}{The first year for the account.}

\item{time_end}{The last year for the account.}
}
\value{
A
\code{\link[dembase:DemographicAccount-class]{DemographicAccount}}.
}
\description{
Make the initial values for a demographic account
consisting of population, deaths, and births
}
\details{
The account will not typically be consistent -
that is, the demographic accounting equations will
not typically be satisfied. The account can be
made consistent using function \code{\link[dembase]{makeConsistent}}.

If \code{Lx} and \code{propn_age_fert} use 5-year time
steps, then \code{time_end - time_start} should be a
multiple of 5.
}
\examples{
Lx <- dembase::Values(Lx_west[ , , 20])
mort_rates <- dembase::Values(mx_west[ , , 20])
propn_age_fert <- dembase::Values(propn_age_fert_booth)
expected_popn <- make_expected_popn(popn_size = 100,
                                    Lx = Lx,
                                    sex_ratio = 105)
fert_rates <- make_stationary_fert_rates(Lx = Lx,
                               propn_age_fert = propn_age_fert,
                               sex_ratio = 105)
account <- make_account(expected_popn = expected_popn,
                        mort_rates = mort_rates,
                        fert_rates = fert_rates,
                        time_start = 1000,
                        time_end = 1020)
dembase::summary(account)
}
