% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_expected_population.R
\name{make_expected_popn}
\alias{make_expected_popn}
\title{Calculate an implied stationary population}
\usage{
make_expected_popn(popn_size, Lx, sex_ratio)
}
\arguments{
\item{popn_size}{The total population size.}

\item{Lx}{An array.}

\item{sex_ratio}{The number of males per 100 females.}
}
\value{
A \code{\link[dembase]{Values}} object,
with the same dimensions and dimnames as \code{Lx}.
}
\description{
Calculation the stationary (ie zero growth rate) population
implied by a given combination of total population size,
life table Lx, and sex ratio.
}
\details{
\code{Lx} must have a dimension called \code{"age"}
and a dimension called \code{"sex"}. The labels for the
age dimension must have the format expected for age
intervals by function \code{\link[dembase]{Counts}}.
The labels for the sex dimension must be \code{"Female"}
and \code{"Male"}.
}
\examples{
Lx <- Lx_west[ , , "10"]
make_expected_popn(popn_size = 100,
                   Lx = Lx,
                   sex_ratio = 105)
}
