% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stable_popn.R
\name{make_stable_popn}
\alias{make_stable_popn}
\title{Calculate an implied stable population}
\usage{
make_stable_popn(popn_size, Lx, fert_rates, sex_ratio, time_start = NULL,
  time_end = NULL)
}
\arguments{
\item{popn_size}{The total population size.}

\item{Lx}{A \code{\link[dembase:Counts-class]{Counts}}
array with the life table Lx values.}

\item{fert_rates}{Age-sex specific fertility rates}

\item{sex_ratio}{The number of male births per 100 female
births.}

\item{time_start}{First time point. Optional.}

\item{time_end}{Last time point. Optional.}
}
\value{
A \code{\link[dembase]{Counts}} object,
with the same dimensions as \code{Lx}.
}
\description{
Calculation the stable (ie constant age structure) population
implied by a given combination of total population size,
life table Lx, age-specific fertility rates, and sex ratio.
}
\details{
\code{Lx} must be a \code{\link[dembase:Counts-class]{Counts}}
array with two dimensions. The dimensions must have
\code{\link[dembase]{dimtype}} \code{"age"} and \code{"sex"}.
The \code{"age"} dimension must have \code{\link[dembase]{dimscale}}
\code{"Intervals"}.

If \code{time_start} and \code{time_end} arguments
are supplied, a time dimension is added to the
result.
}
\examples{
Lx <- dembase::Counts(Lx_west[ , , "10"])
propn_age_fert <- dembase::Values(propn_age_fert_booth)
fert_rates <- make_tfr_fert_rates(tfr = 5,
                                  propn_age_fert = propn_age_fert,
                                  sex_ratio = 105)
make_stable_popn(popn_size = 100,
                 Lx = Lx,
                 sex_ratio = 105,
                 fert_rates = fert_rates)
x <- make_stable_popn(popn_size = 100,
                 Lx = Lx,
                 sex_ratio = 105,
                 fert_rates = fert_rates,
                 time_start = 2000,
                 time_end = 2025)
}
\seealso{
\code{\link{Lx_west}},
\code{\link{make_stationary_popn}},
\code{\link{make_tfr_fert_rates}}
}
