% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stationary_fert_rates.R
\name{make_stationary_fert_rates}
\alias{make_stationary_fert_rates}
\title{Calculate the fertility rates that would produce a
stationary population}
\usage{
make_stationary_fert_rates(Lx, propn_age_fert, sex_ratio)
}
\arguments{
\item{Lx}{A \code{\link[dembase:Values-class]{Values}}
array with the life table Lx values.}

\item{propn_age_fert}{A \code{\link[dembase:Values-class]{Values}}
with the age-distribution of fertility rates.}

\item{sex_ratio}{The number of male births per 100 female
births.}
}
\value{
A \code{\link[dembase]{Values}} object with
\code{"age"} and \code{"sex"} dimensions.
}
\description{
Given a set of Lx, and an age pattern for fertility,
calculate the age-sex-specific fertility rates that would
produce a stationary (ie zero growth rate) population.
}
\details{
\code{Lx} must be a \code{\link[dembase:Values-class]{Values}}
array with two dimensions. The dimensions of \code{Lx} must have
\code{\link[dembase]{dimtype}} \code{"age"} and \code{"sex"}.
The \code{"age"} dimension must have \code{\link[dembase]{dimscale}}
\code{"Intervals"}.

\code{propn_age_fert} must be a
\code{\link[dembase:Values-class]{Values}}
array, with a single dimension. The dimension must have
\code{\link[dembase]{dimtype}} \code{"age"},
and \code{\link[dembase]{dimscale}} \code{"Intervals"}.
}
\examples{
Lx <- dembase::Values(Lx_west[ , , "10"])
propn_age_fert <- dembase::Values(propn_age_fert_booth)
make_stationary_fert_rates(Lx = Lx,
                           propn_age_fert = propn_age_fert,
                           sex_ratio = 105)
}
\seealso{
\code{\link{Lx_west}}, \code{\link{propn_age_fert_booth}},
\code{\link{make_stationary_fert_rates}}
}
