% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stationary_popn.R
\name{make_stationary_popn}
\alias{make_stationary_popn}
\title{Calculate an implied stationary population}
\usage{
make_stationary_popn(
  popn_size,
  Lx,
  sex_ratio,
  time_start = NULL,
  time_end = NULL
)
}
\arguments{
\item{popn_size}{The total population size.}

\item{Lx}{A \code{\link[dembase:Counts-class]{Counts}}
array with the life table Lx values.}

\item{sex_ratio}{The number of male births per 100 female
births.}

\item{time_start}{First time point. Optional.}

\item{time_end}{Last time point. Optional.}
}
\value{
A \code{\link[dembase]{Counts}} object,
with the same dimensions as \code{Lx}.
}
\description{
Calculation the stationary (ie zero growth rate) population
implied by a given combination of total population size,
life table Lx, and sex ratio.
}
\details{
\code{Lx} must be a \code{\link[dembase:Counts-class]{Counts}}
array with two dimensions. The dimensions must have
\code{\link[dembase]{dimtype}} \code{"age"} and \code{"sex"}.
The \code{"age"} dimension must have \code{\link[dembase]{dimscale}}
\code{"Intervals"}.

If \code{time_start} and \code{time_end} arguments
are supplied, a time dimension is added to the
result.
}
\examples{
Lx <- dembase::Counts(Lx_west[ , , "10"])
make_stationary_popn(popn_size = 100,
                     Lx = Lx,
                     sex_ratio = 105)
make_stationary_popn(popn_size = 100,
                     Lx = Lx,
                     sex_ratio = 105,
                     time_start = 2000,
                     time_end = 2010)
}
\seealso{
\code{\link{Lx_west}},
\code{\link{make_stationary_fert_rates}}
}
