% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tfr_fert_rates.R
\name{make_tfr_fert_rates}
\alias{make_tfr_fert_rates}
\title{Calculate the age-specific fertility rates that
produce a given total fertility rate}
\usage{
make_tfr_fert_rates(tfr, propn_age_fert, sex_ratio)
}
\arguments{
\item{tfr}{The target total fertility rate. A positive number.}

\item{propn_age_fert}{A \code{\link[dembase:Values-class]{Values}}
with the age-distribution of fertility rates.}

\item{sex_ratio}{The number of male births per 100 female
births.}
}
\value{
A \code{\link[dembase]{Values}} object with
\code{"age"} and \code{"sex"} dimensions.
}
\description{
Given an age pattern for fertility and a sex ratio at birth,
calculate the age-sex-specific fertility rates that would
yield a target total fertility rate (TFR).
}
\details{
\code{propn_age_fert} must be a
\code{\link[dembase:Counts-class]{Counts}}
array, with a single dimension. The dimension must have
\code{\link[dembase]{dimtype}} \code{"age"},
and \code{\link[dembase]{dimscale}} \code{"Intervals"}.
}
\examples{
propn_age_fert <- dembase::Values(propn_age_fert_booth)
make_tfr_fert_rates(tfr = 5,
                    propn_age_fert = propn_age_fert,
                    sex_ratio = 105)
}
\seealso{
\code{\link{propn_age_fert_booth}},
}
