% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propn_age_fert_booth.R
\docType{data}
\name{propn_age_fert_booth}
\alias{propn_age_fert_booth}
\title{Age distribution for fertility}
\format{An object of class \code{array} of length 8.}
\source{
Booth H (1984) Transforming Gompertz's function
for fertility analysis: The development of a standard
for the relational Gompertz function.
\emph{Population Studies} 38(3): 495-506.
}
\usage{
propn_age_fert_booth
}
\description{
The proportional distribution of fertility by age,
implied by the 'Booth standard'.
}
\details{
The key input is 'Yx' from table 2 of Booth (1984).
}
\keyword{datasets}
