% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propn_age_fert_maori.R
\docType{data}
\name{propn_age_fert_maori}
\alias{propn_age_fert_maori}
\title{Age distribution for fertility, based on Maori fertility in 1962}
\format{
An object of class \code{array} of length 8.
}
\source{
Table "Age-specific fertility rates by 5 year age group
(Maori and total population) (Annual-Dec)",
from the Infoshare database on the Statistics New Zealand
website. Accessed 24 August 2020.
}
\usage{
propn_age_fert_maori
}
\description{
The proportional distribution of fertility by age,
calculated from registered Maori births in 1962,
}
\keyword{datasets}
