% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_expected_popn.R
\name{make_expected_popn}
\alias{make_expected_popn}
\title{Calculate an implied stationary population}
\usage{
make_expected_popn(popn_size, Lx, sex_ratio)
}
\arguments{
\item{popn_size}{The total population size.}

\item{Lx}{A \code{\link[dembase:Values-class]{Values}}
array with the life table Lx values.}

\item{sex_ratio}{The number of male births per 100 female
births.}
}
\value{
A \code{\link[dembase]{Values}} object,
with the same dimensions as \code{Lx}.
}
\description{
Calculation the stationary (ie zero growth rate) population
implied by a given combination of total population size,
life table Lx, and sex ratio.
}
\details{
\code{Lx} must be a \code{\link[dembase:Values-class]{Values}}
array with two dimensions. The dimensions must have
\code{\link[dembase]{dimtype}} \code{"age"} and \code{"sex"}.
The \code{"age"} dimension must have \code{\link[dembase]{dimscale}}
\code{"Intervals"}.
}
\examples{
Lx <- dembase::Values(Lx_west[ , , "10"])
make_expected_popn(popn_size = 100,
                   Lx = Lx,
                   sex_ratio = 105)
}
\seealso{
\code{\link{Lx_west}},
\code{\link{make_stationary_fert_rates}}
}
