% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_west.R
\docType{data}
\name{mx_west}
\alias{mx_west}
\title{Mortality rates mx from the West model life table}
\format{An object of class \code{array} of dimension 17 x 2 x 2 x 25.}
\source{
The values were obtained from the
\code{\link[demogR]{cdmltw}} function
in package \code{demogR}.
}
\usage{
mx_west
}
\description{
Mortality rates mx, by age, sex, Lexis triangle, and life table
level, for the West model life table.
}
\details{
For a description of life tables, and West model life tables,
see \url{http://demographicestimation.iussp.org}.

Above age 5, values for the upper Lexis triangle
are identical to values for the lower Lexis triangle,
for the same age, sex, and lifetable level.
Below age 5, the triangles are weighted
averages of mortality rates for ages 0 and 1-4.
}
\seealso{
\code{\link{Lx_west}}
}
\keyword{datasets}
